refactor(hook): useLoggedInUser improvements
