+++ components/contribution-flow/StepProfile.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+import useLoggedInUser from '../../lib/hooks/useLoggedInUser';
+
 import { Box } from '../Grid';
-import { useLoggedInUser } from '../UserProvider';
 
 import StepProfileGuestForm from './StepProfileGuestForm';
 import StepProfileLoggedInForm from './StepProfileLoggedInForm';
