+++ components/help-and-support/ContactSection.js
@@ -9,6 +9,7 @@ import { isURL } from 'validator';
 import { sendContactMessage } from '../../lib/api';
 import { createError, ERROR, i18nGraphqlException } from '../../lib/errors';
 import { formatFormErrorMessage } from '../../lib/form-utils';
+import useLoggedInUser from '../../lib/hooks/useLoggedInUser';
 import { isValidEmail } from '../../lib/utils';
 
 import Container from '../Container';
@@ -22,7 +23,6 @@ import StyledInput from '../StyledInput';
 import StyledInputField from '../StyledInputField';
 import StyledInputGroup from '../StyledInputGroup';
 import { P, Span } from '../Text';
-import { useLoggedInUser } from '../UserProvider';
 
 const validate = values => {
   const errors = {};
