+++ components/orders/OrdersWithData.js
@@ -8,6 +8,7 @@ import { FormattedMessage } from 'react-intl';
 import { ORDER_STATUS } from '../../lib/constants/order-status';
 import { parseDateInterval } from '../../lib/date-utils';
 import { API_V2_CONTEXT, gqlV2 } from '../../lib/graphql/helpers';
+import useLoggedInUser from '../../lib/hooks/useLoggedInUser';
 import { usePrevious } from '../../lib/hooks/usePrevious';
 
 import { parseAmountRange } from '../budget/filters/AmountFilter';
@@ -20,7 +21,6 @@ import Pagination from '../Pagination';
 import SearchBar from '../SearchBar';
 import StyledHr from '../StyledHr';
 import { H1 } from '../Text';
-import { useLoggedInUser } from '../UserProvider';
 
 import OrdersFilters from './OrdersFilters';
 import OrdersList from './OrdersList';
