+++ lib/hooks/useLoggedInUser.js
@@ -1,0 +1,3 @@
+import { useLoggedInUser } from '../../components/UserProvider';
+
+export default useLoggedInUser;
